updated_bot/requirements.txt
aiogram>=3.4.0
supabase>=2.3.0
python-dotenv>=1.0.0
flask>=3.0.0
httpx>=0.25.0
